fix(napi): address 4 code-quality issues from Phase 4f Task A review
Issue 1 – catch_unwind_napi dead at FFI boundary: add async_catch_unwind_napi
helper in panic.rs (shared panic_payload_message extractor), wrap every sync
napi method in catch_unwind_napi and every async method in
async_catch_unwind_napi so panics are converted to typed RUST_PANIC errors
instead of unwinding across the FFI boundary.

Issue 2 – init_uta silently overwrites duplicates: check for existing account_id
before spawning and return InvalidArg error if already initialized.

Issue 3 – timestamp_ms i64 becomes JS BigInt in napi-rs v2: change field to f64
(JS number, safely represents all ms timestamps past year 2100) and update
EventDispatcher::emit cast accordingly.

Issue 4 – _join handle dropped making actor panics invisible: store both
actor_join and forwarder_join in UtaProxyHandle as Option<JoinHandle<()>>;
abort both in shutdown_uta.

Minor: fix misleading "silently dropped" comment in Task 2 forwarder — events
arriving before subscribe_events are retained in the ring buffer for backfill.

+++ crates/alice-trading-core/src/napi_binding/events.rs
@@ -61,7 +61,7 @@ impl EventDispatcher {
         let event = TradingCoreEvent {
             uta_id: uta_id.to_string(),
             seq,
-            timestamp_ms: chrono::Utc::now().timestamp_millis(),
+            timestamp_ms: chrono::Utc::now().timestamp_millis() as f64,
             event_type: event_type.to_string(),
             payload_json,
         };
